+++ contracts/yNFTVaults/QuickswapYNFTVault.sol
@@ -18,6 +18,19 @@ contract QuickswapYNFTVault is YNFTVault {
   IERC20 public immutable dQuick;
   IERC20 public immutable wMatic;
 
+  event YNftWithdrawn(
+    address pair,
+    uint256 tokenId,
+    uint256 liquidityWithdrawn,
+    uint256 performanceFee
+  );
+  event YNftCreated(address pair, uint256 tokenId, uint256 liquidity);
+  event YNftAssetDeposited(address token, uint256 amount);
+  event YNftLpMiningRewardsAccrued(
+    uint256 dQuickBalance,
+    uint256 wMaticBalance
+  );
+
   constructor(
     IUniswapV2Router02 _dexRouter,
     IUniswapV2Pair _pair,
@@ -72,6 +85,7 @@ contract QuickswapYNFTVault is YNFTVault {
     );
 
     _farmLiquidity(liquidity);
+    emit YNftAssetDeposited(_tokenIn, liquidity);
   }
 
   function depositETH(
@@ -95,6 +109,7 @@ contract QuickswapYNFTVault is YNFTVault {
     );
 
     _farmLiquidity(liquidity);
+    emit YNftAssetDeposited(address(0), liquidity);
   }
 
   function getRewardLPMining() external onlyRole(HARVESTER_ROLE) whenNotPaused {
@@ -103,6 +118,7 @@ contract QuickswapYNFTVault is YNFTVault {
     dQuick.transfer(beneficiary, dQuickBalance);
     uint256 wMaticBalance = wMatic.balanceOf(address(this));
     wMatic.transfer(beneficiary, wMaticBalance);
+    emit YNftLpMiningRewardsAccrued(dQuickBalance, wMaticBalance);
   }
 
   function _withrdrawFromLPMining(uint256 _balance) private {
@@ -132,6 +148,7 @@ contract QuickswapYNFTVault is YNFTVault {
       balanceOf[tokenId] = balance;
       totalSupply = totalSupply + balance;
     }
+    emit YNftCreated(address(pair), tokenId, _liquidity);
     return tokenId;
   }
 
@@ -372,6 +389,14 @@ contract QuickswapYNFTVault is YNFTVault {
     );
 
     yNFT.burn(_nftTokenId);
+    uint256 performanceFeeLiquidity = (balanceToWithdraw *
+      performanceFeeToWithdrawPerMille) / 1000;
+    emit YNftWithdrawn(
+      address(pair),
+      _nftTokenId,
+      balanceToWithdraw,
+      performanceFeeLiquidity
+    );
   }
 
   function withdrawToUnderlyingTokens(
@@ -447,6 +472,12 @@ contract QuickswapYNFTVault is YNFTVault {
     }
 
     yNFT.burn(_nftTokenId);
+    emit YNftWithdrawn(
+      address(pair),
+      _nftTokenId,
+      balanceToWithdraw,
+      performanceFee
+    );
   }
 
   function saveBalancesAtBuyForTokenId(uint256 tokenId) private {