+++ scripts/defender-autotasks/aave-harvester/src/config.ts
@@ -1,3 +1,5 @@
+import { getDeployedVaultsAave } from "../../../../consts/deployed";
+
 export enum ChainIds {
   KOVAN = 42,
   MATIC = 137,
@@ -18,10 +20,15 @@ export const ADDRESSES: NetworksAddresses = {
   },
 };
 
-// TODO: how to handle vault address for diffrent vaults
-export const VAULT_ADDRESS = "0x0bf97F8f2dF1cEEC3EF21827E00F11011aE2d118";
 export const PRICE_FEED_DECIMALS = 8;
 export const SLIPPAGE_PERCANTAGE = 3;
 export const ALL_PERCANTAGE = 100;
 export const DEADLINE_SECONDS = 60;
 export const MIN_CLAIM_AMOUNT = 500000000000000000; // 0.5 MATIC
+export const VAULTS = Object.entries(getDeployedVaultsAave()).map(
+  ([vaultName, vault]) => ({
+    vaultName,
+    vaultAddress: vault.vault,
+    ynftAddress: vault.ynft,
+  })
+);