+++ test/QuickswapYNFTVaultTest.ts
@@ -118,6 +118,117 @@ describe("QuickswapYNFTVault", () => {
     );
   });
 
+  it("createYNFT for underlying tokens", async () => {
+    const amountIn = 1000;
+    const amountOutMinFirstToken = 900;
+    const amountOutMinSecondToken = 800;
+    const amountMinLiqudityFirstToken = 100;
+    const amountMinLiquditySecondToken = 100;
+
+    const AMOUNT_AFTER_SWAP1 = 300;
+
+    await token0Mock.transferFrom.returns(true);
+    await token0Mock.approve.returns(true);
+    uniswapRouterMock.swapExactTokensForTokens.returns([
+      amountIn,
+      AMOUNT_AFTER_SWAP1,
+    ]);
+    await token0Mock.approve.returns(true);
+    await token1Mock.approve.returns(true);
+    await tokenIn.approve.returns(true);
+    uniswapPairMock.approve.returns(true);
+
+    await quickswapYnftVault.createYNFT(
+      token0Mock.address,
+      amountIn,
+      amountOutMinFirstToken,
+      amountOutMinSecondToken,
+      amountMinLiqudityFirstToken,
+      amountMinLiquditySecondToken,
+      DEADLINE
+    );
+
+    expect(uniswapRouterMock.swapExactTokensForTokens).to.have.callCount(1);
+    expect(uniswapRouterMock.swapExactTokensForTokens).to.have.been.calledWith(
+      497,
+      amountOutMinSecondToken,
+      [token0Mock.address, token1Mock.address],
+      quickswapYnftVault.address,
+      DEADLINE
+    );
+
+    await token1Mock.transferFrom.returns(true);
+
+    stakingDualRewardsMock.balanceOf.returns(LIQUIDITY);
+
+    await quickswapYnftVault.createYNFT(
+      token1Mock.address,
+      amountIn,
+      amountOutMinFirstToken,
+      amountOutMinSecondToken,
+      amountMinLiqudityFirstToken,
+      amountMinLiquditySecondToken,
+      DEADLINE
+    );
+    expect(uniswapRouterMock.swapExactTokensForTokens).to.have.callCount(2);
+    expect(uniswapRouterMock.swapExactTokensForTokens).to.have.been.calledWith(
+      497,
+      amountOutMinFirstToken,
+      [token1Mock.address, token0Mock.address],
+      quickswapYnftVault.address,
+      DEADLINE
+    );
+  });
+
+  it("createYNFT should revert if underlying token not approved", async () => {
+    const amountIn = 1000;
+    const amountOutMinFirstToken = 900;
+    const amountOutMinSecondToken = 800;
+    const amountMinLiqudityFirstToken = 100;
+    const amountMinLiquditySecondToken = 100;
+    const AMOUNT_AFTER_SWAP1 = 300;
+
+    await tokenIn.transferFrom.returns(true);
+    await tokenIn.approve.returns(true);
+    uniswapRouterMock.swapExactTokensForTokens.returns([
+      amountIn,
+      AMOUNT_AFTER_SWAP1,
+    ]);
+    ``;
+
+    await token0Mock.approve.returns(false);
+    await token1Mock.approve.returns(true);
+
+    await expectRevert(
+      quickswapYnftVault.createYNFT(
+        tokenIn.address,
+        amountIn,
+        amountOutMinFirstToken,
+        amountOutMinSecondToken,
+        amountMinLiqudityFirstToken,
+        amountMinLiquditySecondToken,
+        DEADLINE
+      ),
+      "approve failed."
+    );
+
+    await token0Mock.approve.returns(true);
+    await token1Mock.approve.returns(false);
+
+    await expectRevert(
+      quickswapYnftVault.createYNFT(
+        tokenIn.address,
+        amountIn,
+        amountOutMinFirstToken,
+        amountOutMinSecondToken,
+        amountMinLiqudityFirstToken,
+        amountMinLiquditySecondToken,
+        DEADLINE
+      ),
+      "approve failed."
+    );
+  });
+
   it("createYNFT should call swapExactTokensForTokens for both tokens", async () => {
     const amountIn = 1000;
     const amountOutMinFirstToken = 900;
@@ -165,6 +276,49 @@ describe("QuickswapYNFTVault", () => {
     );
   });
 
+  it("createYNFTForEther should revert if one of tokens not approved", async () => {
+    const amountIn = 1000;
+    const amountOutMinFirstToken = 900;
+    const amountOutMinSecondToken = 800;
+    const amountMinLiqudityFirstToken = 100;
+    const amountMinLiquditySecondToken = 100;
+
+    const AMOUNT_AFTER_SWAP1 = 300;
+
+    uniswapRouterMock.swapExactETHForTokens.returns([
+      amountIn,
+      AMOUNT_AFTER_SWAP1,
+    ]);
+
+    await token1Mock.approve.returns(false);
+    uniswapPairMock.approve.returns(true);
+
+    await expectRevert(
+      quickswapYnftVault.createYNFTForEther(
+        amountOutMinFirstToken,
+        amountOutMinSecondToken,
+        amountMinLiqudityFirstToken,
+        amountMinLiquditySecondToken,
+        DEADLINE
+      ),
+      "approve failed."
+    );
+
+    await token0Mock.approve.returns(false);
+    await token1Mock.approve.returns(true);
+
+    await expectRevert(
+      quickswapYnftVault.createYNFTForEther(
+        amountOutMinFirstToken,
+        amountOutMinSecondToken,
+        amountMinLiqudityFirstToken,
+        amountMinLiquditySecondToken,
+        DEADLINE
+      ),
+      "approve failed."
+    );
+  });
+
   it("createYNFTForEther should call swapExactETHForTokens for both tokens", async () => {
     const amountIn = 1000;
     const amountOutMinFirstToken = 900;
@@ -205,6 +359,43 @@ describe("QuickswapYNFTVault", () => {
     );
   });
 
+  it("createYNFTForEther should omit one swap if WMATIC in pair", async () => {
+    const amountIn = 1000;
+    const amountOutMinFirstToken = 900;
+    const amountOutMinSecondToken = 800;
+    const amountMinLiqudityFirstToken = 100;
+    const amountMinLiquditySecondToken = 100;
+
+    const AMOUNT_AFTER_SWAP1 = 300;
+
+    uniswapRouterMock.swapExactETHForTokens.returns([
+      amountIn,
+      AMOUNT_AFTER_SWAP1,
+    ]);
+
+    uniswapRouterMock.WETH.returns(token0Mock.address);
+
+    await token0Mock.approve.returns(true);
+    await token1Mock.approve.returns(true);
+    uniswapPairMock.approve.returns(true);
+
+    await quickswapYnftVault.createYNFTForEther(
+      amountOutMinFirstToken,
+      amountOutMinSecondToken,
+      amountMinLiqudityFirstToken,
+      amountMinLiquditySecondToken,
+      DEADLINE
+    );
+
+    expect(uniswapRouterMock.swapExactETHForTokens).to.have.callCount(1);
+    expect(uniswapRouterMock.swapExactETHForTokens).to.have.been.calledWith(
+      amountOutMinSecondToken,
+      [token0Mock.address, token1Mock.address],
+      quickswapYnftVault.address,
+      DEADLINE
+    );
+  });
+
   it("withdrawToUnderlyingTokens should call removeLiquidity", async () => {
     uniswapRouterMock.swapExactETHForTokens.returns([1000, 300]);
     await token0Mock.approve.returns(true);
@@ -231,6 +422,44 @@ describe("QuickswapYNFTVault", () => {
       LIQUIDITY,
       900,
       800,
+      signers[0].address,
+      DEADLINE
+    );
+  });
+
+  it("withdrawToUnderlyingTokens should call removeLiquidityETH when WMATIC in pair", async () => {
+    uniswapRouterMock.swapExactETHForTokens.returns([1000, 300]);
+    await token0Mock.approve.returns(true);
+    await token1Mock.approve.returns(true);
+    uniswapPairMock.approve.returns(true);
+    stakingDualRewardsMock.balanceOf.returns(LIQUIDITY);
+
+    uniswapRouterMock.addLiquidityETH.returns([
+      LIQUIDITY,
+      LIQUIDITY,
+      LIQUIDITY,
+    ]);
+    await uniswapRouterMock.WETH.returns(token0Mock.address);
+
+    await quickswapYnftVault.createYNFTForEther(900, 800, 100, 100, 101);
+
+    const TOKEN_ID = 0;
+    const amountOutMinFirstToken = 900;
+    const amountOutMinSecondToken = 800;
+
+    await quickswapYnftVault.withdrawToUnderlyingTokens(
+      TOKEN_ID,
+      amountOutMinFirstToken,
+      amountOutMinSecondToken,
+      DEADLINE
+    );
+
+    expect(uniswapRouterMock.removeLiquidityETH).to.have.callCount(1);
+    expect(uniswapRouterMock.removeLiquidityETH).to.have.been.calledWith(
+      token1Mock.address,
+      LIQUIDITY,
+      800,
+      900,
       signers[0].address,
       DEADLINE
     );
@@ -313,6 +542,56 @@ describe("QuickswapYNFTVault", () => {
     );
   });
 
+  it("withdrawToEther should call removeLiquidityETH if WMATIC in pair", async () => {
+    uniswapRouterMock.swapExactETHForTokens.returns([1000, 300]);
+    await token0Mock.approve.returns(true);
+    await token1Mock.approve.returns(true);
+    uniswapPairMock.approve.returns(true);
+    stakingDualRewardsMock.balanceOf.returns(LIQUIDITY);
+    uniswapRouterMock.addLiquidityETH.returns([
+      LIQUIDITY,
+      LIQUIDITY,
+      LIQUIDITY,
+    ]);
+    await uniswapRouterMock.WETH.returns(token0Mock.address);
+    await quickswapYnftVault.createYNFTForEther(900, 800, 100, 100, 101);
+
+    const TOKEN_ID = 0;
+    const amountOutMinFirstToken = 900;
+    const amountOutMinSecondToken = 800;
+    const amountOfEthMin = 1000;
+
+    await uniswapRouterMock.swapExactTokensForETH.returns([
+      amountOutMinFirstToken,
+      amountOfEthMin,
+    ]);
+
+    await quickswapYnftVault.withdrawToEther(
+      TOKEN_ID,
+      amountOutMinFirstToken,
+      amountOutMinSecondToken,
+      amountOfEthMin,
+      DEADLINE
+    );
+
+    expect(uniswapRouterMock.removeLiquidityETH).to.have.callCount(1);
+    expect(uniswapRouterMock.removeLiquidityETH).to.have.been.calledWith(
+      token1Mock.address,
+      LIQUIDITY,
+      800,
+      900,
+      quickswapYnftVault.address,
+      DEADLINE
+    );
+  });
+
+  it("should not allow performance fee higher than 20%", async () => {
+    await expectRevert(
+      quickswapYnftVault.setPerformanceFee(201),
+      "Performance Fee cannot be that much"
+    );
+  });
+
   it("depositTokens should swap and add liquidity", async () => {
     const amountIn = 1000;
     const amountOutMinFirstToken = 900;
@@ -371,6 +650,25 @@ describe("QuickswapYNFTVault", () => {
       DEADLINE
     );
     expect(stakingDualRewardsMock.stake).to.have.been.calledWith(LIQUIDITY);
+  });
+
+  it("depositTokens should revert if LP tokens deposited", async () => {
+    await quickswapYnftVault.grantRole(
+      ethers.utils.keccak256(ethers.utils.toUtf8Bytes("HARVESTER_ROLE")),
+      signers[0].address
+    );
+
+    await expectRevert(
+      quickswapYnftVault.depositTokens(
+        uniswapPairMock.address,
+        0,
+        0,
+        0,
+        0,
+        DEADLINE
+      ),
+      "Cannot deposit LP tokens"
+    );
   });
 
   it("depositETH should swap and add liquidity", async () => {
@@ -656,6 +954,43 @@ describe("QuickswapYNFTVault", () => {
     );
 
     expect(uniswapRouterMock.removeLiquidity).to.have.callCount(4);
+  });
+
+  it("should calculate correct performance fee on withdrawToUnderlyingTokens if WMATIC in pair", async () => {
+    const EXTRACTED_TOKEN_ID = 0;
+    const LIQUIDITY_REWARDS = 111;
+    const EXPECTED_PAYOUT = 508;
+    const PERFORMANCE_FEE = 17;
+    await uniswapRouterMock.WETH.returns(token0Mock.address);
+    uniswapRouterMock.addLiquidityETH.returns([
+      LIQUIDITY,
+      LIQUIDITY,
+      LIQUIDITY,
+    ]);
+
+    await createTwoYNfts(LIQUIDITY_REWARDS, signers, DEADLINE);
+
+    await quickswapYnftVault
+      .connect(signers[1])
+      .withdrawToUnderlyingTokens(EXTRACTED_TOKEN_ID, 0, 0, DEADLINE);
+
+    expect(uniswapRouterMock.removeLiquidityETH).to.have.callCount(2);
+    expect(uniswapRouterMock.removeLiquidityETH).to.have.been.calledWith(
+      token1Mock.address,
+      EXPECTED_PAYOUT - PERFORMANCE_FEE,
+      0,
+      0,
+      signers[1].address,
+      DEADLINE
+    );
+    expect(uniswapRouterMock.removeLiquidityETH).to.have.been.calledWith(
+      token1Mock.address,
+      PERFORMANCE_FEE,
+      0,
+      0,
+      signers[0].address,
+      DEADLINE
+    );
   });
 
   it("should calculate correct performance fee on withdrawToUnderlyingTokens when withdrawing second yNFT", async () => {